requirements.txt
Flask
flask-cors
pandas
openpyxl
gunicorn
PyJWT